requirements.txt
streamlit
torch
torchvision
opencv-python
fer
pillow
numpy
google-api-python-client
lyricsgenius
openai-clip
isodate